dev-requirements.txt
autopep8==1.6.0
django-debug-toolbar==3.2.2
django-extensions==3.1.5
django-stubs==1.9.0
django-stubs-ext==0.3.1
mypy==0.910
mypy-extensions==0.4.3
pycodestyle==2.8.0
toml==0.10.2
types-pytz==2021.3.1
types-PyYAML==6.0.1
typing_extensions==4.0.0
